Test trace id in E2E tests
