+++ e2e-tests/dragent_tests/test_streaming.py
@@ -17,12 +17,12 @@ from __future__ import annotations
 import httpx
 from datarobot_genai.core.agents.verify import validate_sequence
 
-from dragent_tests.helpers import assert_tracing_conventions
 from dragent_tests.helpers import collect_ag_ui_events
 from dragent_tests.helpers import collect_text
 from dragent_tests.helpers import make_generate_payload
 from dragent_tests.helpers import stream_sse_responses
-from dragent_tests.mock_otel_collector import MockOtelCollector
+from dragent_tests.otel_helpers import MockOtelCollector
+from dragent_tests.otel_helpers import assert_tracing_conventions
 
 
 def test_generate_streaming(
